xtask: add profiling tool
Add two commands (profile and exec) to xtask to enable executing `samply` over a custom workload

+++ xtask/src/nomt.rs
@@ -12,8 +12,11 @@ pub struct NomtDB {
 }
 
 impl NomtDB {
-    pub fn new() -> Self {
-        let _ = std::fs::remove_dir_all("nomt_db");
+    pub fn new(reset: bool) -> Self {
+        if reset {
+            // Delete previously existing db
+            let _ = std::fs::remove_dir_all("nomt_db");
+        }
 
         let opts = Options {
             path: PathBuf::from("nomt_db"),
